refactor: 패키지 분리
- controller 패키지 추가하여 RacingController 할당
- domain/racing 패키지 추가하여 자동차 경주와 관련된 클래스 할당
- car 패키지에 참가자와 횟수 관련된 클래스 할당
