+++ src/main/java/racingcar/domain/car/Turn.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import java.util.regex.Pattern;
 import racingcar.view.UserConsole;
